[BE #2 ] Feature: init project 수정
프로젝트 패키지 이름 example을 codesquad로 수정
swagger setting 을 2.7.0 -> 2.9.2
swagger config 에서 base package 설정
application 파일을 properties 에서 yml로 수정
dev 환경에서는 h2 db를 사용하기 위한 의존성, 설정 추가
