+++ BE/src/main/java/com/codesquad/sidedish08/Sidedish08Application.java
@@ -1,0 +1,12 @@
+package com.codesquad.sidedish08;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Sidedish08Application {
+
+  public static void main(String[] args) {
+    SpringApplication.run(Sidedish08Application.class, args);
+  }
+}
