+++ BE/src/main/java/com/codesquad/sidedish08/msessage/ErrorMessages.java
@@ -1,0 +1,6 @@
+package com.codesquad.sidedish08.msessage;
+
+public class ErrorMessages {
+
+  public static final String ERROR = "ERROR";
+}
