+++ BE/src/main/java/com/codesquad/sidedish08/msessage/SuccessMessages.java
@@ -1,0 +1,6 @@
+package com.codesquad.sidedish08.msessage;
+
+public class SuccessMessages {
+
+  public static final String SUCCESS = "SUCCESS";
+}
