+++ BE/src/main/java/com/codesquad/sidedish08/service/BasicService.java
@@ -1,0 +1,10 @@
+package com.codesquad.sidedish08.service;
+
+import java.util.Map;
+
+public interface BasicService {
+
+  Map<String, ?> dish();
+
+  Map<String, ?> detail(String detailHash);
+}
