+++ BE/src/main/java/com/codesquad/sidedish08/service/BestDishService.java
@@ -1,0 +1,17 @@
+package com.codesquad.sidedish08.service;
+
+import com.codesquad.sidedish08.util.ResponseUtils;
+import java.util.Map;
+import org.springframework.stereotype.Service;
+
+@Service
+public class BestDishService implements BasicService {
+
+  public Map<String, ?> dish() {
+    return ResponseUtils.getResultMap("dish", null);
+  }
+
+  public Map<String, ?> detail(String detailHash) {
+    return ResponseUtils.getResultMap("dish", null);
+  }
+}
