+++ BE/src/main/java/com/codesquad/sidedish08/util/ResponseUtils.java
@@ -1,0 +1,13 @@
+package com.codesquad.sidedish08.util;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class ResponseUtils {
+
+  public static <T> Map<String, T> getResultMap(String key, T value) {
+    Map<String, T> result = new HashMap<>();
+    result.put(key, value);
+    return result;
+  }
+}
