--- BE/src/main/java/com/example/sidedish08/Sidedish08Application.java
@@ -1,12 +1,0 @@
-package com.example.sidedish08;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Sidedish08Application {
-
-  public static void main(String[] args) {
-    SpringApplication.run(Sidedish08Application.class, args);
-  }
-}
