--- BE/src/main/java/com/example/sidedish08/config/SwaggerConfig.java
@@ -1,22 +1,0 @@
-package com.example.sidedish08.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig {
-
-  @Bean
-  public Docket api() {
-    return new Docket(DocumentationType.SWAGGER_2)
-        .groupName("Api")
-        .select()
-        .apis(RequestHandlerSelectors.any())
-        .build();
-  }
-}
