--- BE/src/main/java/com/example/sidedish08/exception/CustomAdvice.java
@@ -1,18 +1,0 @@
-package com.example.sidedish08.exception;
-
-import com.example.sidedish08.msessage.ErrorMessages;
-import com.example.sidedish08.response.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class CustomAdvice {
-
-  @ExceptionHandler(Exception.class)
-  @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-  public ErrorResponse handleException(Exception e) {
-    return new ErrorResponse(ErrorMessages.ERROR, e.getMessage());
-  }
-}
