--- BE/src/main/java/com/example/sidedish08/msessage/ErrorMessages.java
@@ -1,6 +1,0 @@
-package com.example.sidedish08.msessage;
-
-public class ErrorMessages {
-
-  public static final String ERROR = "ERROR";
-}
