--- BE/src/main/java/com/example/sidedish08/msessage/SuccessMessages.java
@@ -1,6 +1,0 @@
-package com.example.sidedish08.msessage;
-
-public class SuccessMessages {
-
-  public static final String SUCCESS = "SUCCESS";
-}
