--- BE/src/main/java/com/example/sidedish08/service/BasicService.java
@@ -1,10 +1,0 @@
-package com.example.sidedish08.service;
-
-import java.util.Map;
-
-public interface BasicService {
-
-  Map<String, ?> dish();
-
-  Map<String, ?> detail(String detailHash);
-}
