--- BE/src/main/java/com/example/sidedish08/util/ResponseUtils.java
@@ -1,13 +1,0 @@
-package com.example.sidedish08.util;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class ResponseUtils {
-
-  public static <T> Map<String, T> getResultMap(String key, T value) {
-    Map<String, T> result = new HashMap<>();
-    result.put(key, value);
-    return result;
-  }
-}
