--- BE/src/main/resources/application.properties
@@ -1,9 +1,0 @@
-# logging level 설정
-logging.level.root=debug
-logging.level.sql=trace
-# log file path
-logging.file.path=./logs
-logging.file.name=${logging.file.path}/dev_log.log
-# logging pattern
-logging.pattern.file=%d{dd-MM-yyyy HH:mm:ss.SSS,Asia/Seoul} [%thread] %-5level %logger{36}.%M - %msg%n
-logging.pattern.console=%d{dd-MM-yyyy HH:mm:ss.SSS,Asia/Seoul} %magenta([%thread]) %highlight(%-5level) %logger.%M - %msg%n
