+++ BE/src/test/java/com/codesquad/sidedish08/Sidedish08ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.codesquad.sidedish08;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Sidedish08ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
