--- BE/src/test/java/com/example/sidedish08/Sidedish08ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.sidedish08;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Sidedish08ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
